+++ swing/common-model/src/test/java/is/codion/swing/common/model/worker/ProgressWorkerTest.java
@@ -36,7 +36,7 @@ public final class ProgressWorkerTest {
 
 		List<Integer> stateChanges = new ArrayList<>();
 
-		ProgressWorker.ProgressTask<Integer, String> task = progressReporter -> {
+		ProgressWorker.ProgressResultTask<Integer, String> task = progressReporter -> {
 			Thread.sleep(100);
 			progressReporter.report(100);
 			progressReporter.publish("Done");